PO mapping is seller + quote => po
