+++ src/contracts/Nethereum.Commerce.Contracts/Contracts/IPoStorage.sol
@@ -14,5 +14,6 @@ interface IPoStorage
     // PO documents
     function setPo(IPoTypes.Po calldata po) external;
     function getPo(uint poNumber) external view returns (IPoTypes.Po memory po);
-    function getPoNumberByApproverAndQuote(address approverAddress, uint quoteId) external view returns (uint poNumber);
+    function getPoNumberBySellerAndQuote(bytes32 sellerId, uint quoteId) external view returns (uint poNumber);
 }
+
